Update dd scripts to only delete test file if it didn't exist previously

+++ lib/blog_snippets/articles/tuning_dd_block_size/dd_ibs_test.sh
@@ -4,6 +4,7 @@
 set -e
 
 TEST_FILE=${1:-dd_ibs_testfile}
+[ -e "$TEST_FILE" ]; TEST_FILE_EXISTS=$?
 TEST_FILE_SIZE=134217728
 
 # Exit if file exists
@@ -34,5 +35,5 @@ do
   printf "$PRINTF_FORMAT" "$BLOCK_SIZE" "$TRANSFER_RATE"
 done
 
-# Clean up
-rm $TEST_FILE
+# Clean up the test file if we created one
+[ $TEST_FILE_EXISTS -ne 0 ] && rm $TEST_FILE
